@@ -3,12 +3,12 @@
   "version": "0.1.0",
   "license": "MIT",
   "engines": {
-    "node": "8.10.0",
-    "npm": "5.6.0"
+    "node": "8.11.3",
+    "yarn": "1.9.4"
   },
   "scripts": {
     "ng": "ng",
-    "start": "ng serve -H 0.0.0.0",
+    "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint --type-check",
@@ -18,39 +18,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/router": "5.2.9",
-    "core-js": "2.5.3",
-    "hint.css": "2.5.0",
-    "rxjs": "5.5.7",
-    "zone.js": "git://github.com/JiaLiPassion/zone.js#jasmine-dist"
+    "@angular/common": "^6.1.2",
+    "@angular/compiler": "^6.1.2",
+    "@angular/core": "^6.1.2",
+    "@angular/forms": "^6.1.2",
+    "@angular/http": "^6.1.2",
+    "@angular/platform-browser": "^6.1.2",
+    "@angular/platform-browser-dynamic": "^6.1.2",
+    "@angular/router": "^6.1.2",
+    "core-js": "^2.5.7",
+    "hint.css": "^2.5.0",
+    "rxjs": "^6.2.2",
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "5.2.9",
-    "@types/jasmine": "2.8.6",
-    "@types/node": "8.9.4",
-    "codelyzer": "4.2.1",
-    "husky": "0.14.3",
-    "jasmine-core": "2.99.1",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "1.1.1",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "prettier": "1.11.1",
-    "pretty-quick": "1.4.1",
-    "protractor": "5.3.0",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "2.8.0-dev.20180318"
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "^6.1.3",
+    "@angular/compiler-cli": "^6.1.2",
+    "@types/jasmine": "^2.8.8",
+    "@types/node": "^8.10.25",
+    "codelyzer": "^4.4.3",
+    "husky": "^0.14.3",
+    "jasmine-core": "^3.2.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^3.0.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-cli": "^1.0.1",
+    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^1.2.0",
+    "prettier": "^1.14.2",
+    "pretty-quick": "^1.6.0",
+    "protractor": "^5.4.0",
+    "ts-node": "^7.0.1",
+    "tslint": "^5.11.0",
+    "typescript": "^2.9.2"
   }
 }
